remove oqm package and related tests
